Modelling Soil Erosion
In this project, we modell the erosion of soil, as it happens when there is a shallow flow of water over a slope. This can for example frequently be observed on beaches. The soils forms a wave -like structure, which will, depending on the Froude number of the flow, travel upstream or downstream. We modell the wave-like structure of the soil and we solve the resulting system of coupled PDEs numerically.